// src/api/generated/model/contestTeamMemberDetail.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */
import type { ContestTeamMemberQuestionStatistics } from './contestTeamMemberQuestionStatistics';
import type { ContestTeamMemberSubmissionStatistics } from './contestTeamMemberSubmissionStatistics';

export interface ContestTeamMemberDetail {
  contest_team_member_id: string;
  user_id: string;
  name: string;
  email: string;
  is_leader?: boolean;
  is_participated?: boolean;
  score?: number;
  started_at?: string | null;
  base_end_time?: string | null;
  ended_at?: string | null;
  extra_time_seconds?: number;
  remaining_time_seconds?: number;
  is_flagged?: boolean;
  flagged_at?: string | null;
  flagged_reason?: string | null;
  submission_statistics?: ContestTeamMemberSubmissionStatistics;
  question_statistics?: ContestTeamMemberQuestionStatistics;
}